Better types in place, although they will need manual updating
